fix: made button invisible & cleaned out files from unused code
