+++ src/components/ChatForm/ChatForm.tsx
@@ -3,11 +3,7 @@ import React, { useCallback } from "react";
 import { Flex, Card, Text } from "@radix-ui/themes";
 import styles from "./ChatForm.module.css";
 
-import {
-  PaperPlaneButton,
-  BackToSideBarButton,
-  ThreadHistoryButton,
-} from "../Buttons/Buttons";
+import { PaperPlaneButton, BackToSideBarButton } from "../Buttons/Buttons";
 import { TextArea, TextAreaProps } from "../TextArea";
 import { Form } from "./Form";
 import { useOnPressedEnter, useIsOnline, useConfig } from "../../hooks";
@@ -28,8 +24,6 @@ import {
   getInformationMessage,
 } from "../../features/Errors/informationSlice";
 import { InformationCallout } from "../Callout/Callout";
-import { push } from "../../features/Pages/pagesSlice";
-import { selectChatId, selectMessages } from "../../features/Chat";
 
 export type ChatFormProps = {
   onSubmit: (str: string) => void;
@@ -74,12 +68,10 @@ export const ChatForm: React.FC<ChatFormProps> = ({
   const information = useAppSelector(getInformationMessage);
   const [helpInfo, setHelpInfo] = React.useState<React.ReactNode | null>(null);
   const onClearError = useCallback(() => dispatch(clearError()), [dispatch]);
-  const chatId = useAppSelector(selectChatId);
   const onClearInformation = useCallback(
     () => dispatch(clearInformation()),
     [dispatch],
   );
-  const messages = useAppSelector(selectMessages);
   const [value, setValue] = React.useState("");
 
   const { checkboxes, onToggleCheckbox, setInteracted, unCheckAll } =
@@ -158,10 +150,6 @@ export const ChatForm: React.FC<ChatFormProps> = ({
     },
     [setInteracted, handleHelpInfo],
   );
-
-  const handleThreadHistoryPage = useCallback(() => {
-    dispatch(push({ name: "thread history page", chatId }));
-  }, [chatId, dispatch]);
 
   if (error) {
     return (
@@ -252,13 +240,7 @@ export const ChatForm: React.FC<ChatFormProps> = ({
                 onClick={onClose}
               />
             )}
-            {messages.length !== 0 && !isStreaming && (
-              <ThreadHistoryButton
-                title="View chat thread history"
-                size="1"
-                onClick={handleThreadHistoryPage}
-              />
-            )}
+            {/* TODO: Reserved space for microphone button coming later on */}
             <PaperPlaneButton
               disabled={isStreaming || !isOnline}
               title="send"
