+++ src/components/Toolbar/Dropdown.tsx
@@ -20,7 +20,6 @@ export type DropdownNavigationOptions =
   | "hot keys"
   | "restart tour"
   | "cloud login"
-  | "chat history"
   | "";
 
 type DropdownProps = {
